refactor(cdc): extract RunInit base-export driver from cdc-init CLI (#173)
Mechanical move of the init driver (context setup, schema iteration,
batch export, manifest update, batch sizing) from cmd/tools/cdc_init.go
into internal/cdc (init.go + init_sizing.go) behind an exported
cdc.RunInit(ctx, InitOptions) (InitSummary, error). The CLI keeps flag
parsing and delegates. No behavior change.
